+++ src/EnrolmentWorkflow/Request/class.RequestsGUI.php
@@ -5,8 +5,11 @@ namespace srag\Plugins\SrUserEnrolment\EnrolmentWorkflow\Request;
 use ilObjCourseGUI;
 use ilRepositoryGUI;
 use ilSrUserEnrolmentPlugin;
+use ilSubmitButton;
 use ilUIPluginRouterGUI;
+use srag\CustomInputGUIs\SrUserEnrolment\MultiSelectSearchInputGUI\MultiSelectSearchInputGUI;
 use srag\DIC\SrUserEnrolment\DICTrait;
+use srag\Plugins\SrUserEnrolment\EnrolmentWorkflow\Step\StepGUI;
 use srag\Plugins\SrUserEnrolment\Utils\SrUserEnrolmentTrait;
 
 /**
@@ -27,6 +30,7 @@ class RequestsGUI
     const CMD_APPLY_FILTER = "applyFilter";
     const CMD_BACK = "back";
     const CMD_GET_USERS_AUTO_COMPLETE = "getUsersAutoComplete";
+    const CMD_GET_USERS_AUTO_COMPLETE_REQUEST = "getUsersAutoCompleteRequest";
     const CMD_LIST_REQUESTS = "listRequests";
     const CMD_RESET_FILTER = "resetFilter";
     const GET_PARAM_REF_ID = "ref_id";
@@ -76,6 +80,7 @@ class RequestsGUI
                     case self::CMD_APPLY_FILTER:
                     case self::CMD_BACK:
                     case self::CMD_GET_USERS_AUTO_COMPLETE:
+                    case self::CMD_GET_USERS_AUTO_COMPLETE_REQUEST:
                     case self::CMD_LIST_REQUESTS:
                     case self::CMD_RESET_FILTER:
                         $this->{$cmd}();
@@ -111,6 +116,23 @@ class RequestsGUI
         if (!empty($this->obj_ref_id)) {
             self::dic()->tabs()->setBackTarget(self::dic()->objDataCache()->lookupTitle(self::dic()->objDataCache()->lookupObjId($this->obj_ref_id)), self::dic()->ctrl()
                 ->getLinkTarget($this, self::CMD_BACK));
+
+            self::dic()->ctrl()->saveParameterByClass(RequestStepGUI::class, self::GET_PARAM_REF_ID);
+            $step = current(self::srUserEnrolment()->enrolmentWorkflow()->steps()->getSteps(self::srUserEnrolment()
+                ->enrolmentWorkflow()
+                ->selectedWorkflows()
+                ->getWorkflowId(self::dic()->objDataCache()->lookupObjId($this->obj_ref_id))));
+            self::dic()->ctrl()->setParameterByClass(RequestStepGUI::class, StepGUI::GET_PARAM_STEP_ID, $step->getStepId());
+            self::dic()->toolbar()->setFormAction(self::dic()->ctrl()->getFormActionByClass(RequestStepGUI::class));
+
+            $users = new MultiSelectSearchInputGUI("", RequestStepGUI::GET_PARAM_USER_ID);
+            $users->setAjaxLink(self::dic()->ctrl()->getLinkTarget($this, self::CMD_GET_USERS_AUTO_COMPLETE_REQUEST, "", true, false));
+            self::dic()->toolbar()->addInputItem($users);
+
+            $request_button = ilSubmitButton::getInstance();
+            $request_button->setCaption($step->getActionTitle(), false);
+            $request_button->setCommand(RequestStepGUI::CMD_REQUEST_STEP);
+            self::dic()->toolbar()->addButtonInstance($request_button);
         }
 
         self::dic()->tabs()->addTab(self::TAB_REQUESTS, self::plugin()->translate("requests", self::LANG_MODULE), self::dic()->ctrl()
@@ -159,6 +181,35 @@ class RequestsGUI
                 "id"   => $id,
                 "text" => $title
             ];
+        }
+
+        self::output()->outputJSON(["result" => $options]);
+    }
+
+
+    /**
+     *
+     */
+    protected function getUsersAutoCompleteRequest()/*: void*/
+    {
+        $search = strval(filter_input(INPUT_GET, "term", FILTER_DEFAULT, FILTER_FORCE_ARRAY)["term"]);
+
+        $options = [];
+
+        foreach (self::srUserEnrolment()->ruleEnrolment()->searchUsers($search) as $id => $title) {
+            if (self::srUserEnrolment()
+                ->enrolmentWorkflow()
+                ->requests()
+                ->canRequestWithAssistant($this->obj_ref_id, current(self::srUserEnrolment()->enrolmentWorkflow()->steps()->getSteps(self::srUserEnrolment()
+                    ->enrolmentWorkflow()
+                    ->selectedWorkflows()
+                    ->getWorkflowId(self::dic()->objDataCache()->lookupObjId($this->obj_ref_id))))->getStepId(), self::dic()->user()->getId(), $id)
+            ) {
+                $options[] = [
+                    "id"   => $id,
+                    "text" => $title
+                ];
+            }
         }
 
         self::output()->outputJSON(["result" => $options]);